requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
fastmcp>=2.0.0
jinja2>=3.1.4
python-multipart>=0.0.12
itsdangerous>=2.2.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.140.0
python-dotenv>=1.0.1
httpx>=0.27.0
pypdf>=5.0.0
Pillow>=10.4.0
pytesseract>=0.3.13
